+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.0.5 
 
+* Add `requestTimeout` to cache configuration. Cache requests that are not fulfilled within this duration are closed with a `StreamRequestTimedOutException`. Defaults to 30 seconds.
+
 * Fix broken links in example project
 
 * Fix potential I/O exceptions